Services restructured, dependencies frozen, output_handler connected
The docker services have been rearrange and renamed to a consistent pattern.
Container dependencies have been added to the docker-compose files.
Synced and froze versions for all socketio, python packages, and
dockerfile base images.
Fixed the output_handler to connect properly to the connector container.
