+++ api/frontend/gui/templates/index.html
@@ -43,7 +43,7 @@
         <h2>Current Jobs</h2>
         <table id="job-list"></table>
 
-        <script type="text/javascript" src="http://cdnjs.cloudflare.com/ajax/libs/socket.io/1.4.5/socket.io.min.js"></script>
+        <script type="text/javascript" src="http://cdnjs.cloudflare.com/ajax/libs/socket.io/3.0.4/socket.io.min.js"></script>
 
         <script type="text/javascript">
             document.addEventListener("DOMContentLoaded", () => {
