+++ api/frontend/gui/wsgi.py
@@ -1,4 +1,4 @@
-from frontend import app
+from gui import app
 
 if __name__ == "__main__":
     app.run()
