+++ api/frontend/output_handler/output_handler.py
@@ -1,4 +1,5 @@
 import os
+import time
 
 import socketio
