+++ api/node/connector/entrypoint.sh
@@ -2,4 +2,4 @@
 
 cd /opt/connector/
 echo "Starting Gunicorn"
-exec gunicorn -k gevent -w 1 --bind 0.0.0.0:6000 --log-level=debug --reload wsgi:app
+exec gunicorn -k gevent -w 1 --bind 0.0.0.0:6000 --log-level=warning --reload wsgi:app
