+++ api/node/connector/requirements.txt
@@ -1,4 +1,4 @@
-flask
-flask-socketio
-gevent
-gunicorn
+flask==1.1.2
+flask-socketio==5.0.0
+gevent==20.9.0
+gunicorn==20.0.4
