+++ api/node/updater/Dockerfile
@@ -1,4 +1,4 @@
-FROM docker:dind
+FROM docker:19.03.14-dind
 
 RUN apk add --no-cache python3 python3-dev py3-pip
