+++ api/node/updater/requirements.txt
@@ -1,1 +1,2 @@
-python-socketio[client]
+python-engineio[client]==4.0.0
+python-socketio[client]==5.0.3
